Update index.html
Some changes to Majula and Things Betwixt.  I will continue working on it some more tomorrow.  Majula will be a bit of a bear because I want to move everything regarding Majula to the beginning so that any player can just look straight there to see what is available, when it is available, and check it off if obtained or unimportant.

Thanks again for your help with this.  I am going very slow not to screw it up, but it is slowly making more sense.